Fix chosen incidents so that clicking a link will not unselect an incident.
Fix chosen incident so that choosing a new list order will either keep the selected incident or unselect it depending on if it exists in the new list.
Fix chosen incident so that choosing a state will either select or unselect the incident depending on if  it exists in the new list.